refactored editor and scroll top button
