+++ src/components/bottomTabBar/view/bottomTabBarStyles.ts
@@ -32,25 +32,4 @@ export default EStyleSheet.create({
   circle: {
     bottom: roundPx(25),
   },
-  fabWrapper: {
-    position: 'absolute',
-    right: roundPx(16),
-    // bottom is overridden from component via { bottom: _fabOffset }
-    zIndex: 20,
-    elevation: 20,
-  },
-
-  fabButton: {
-    width: roundPx(56),
-    height: roundPx(56),
-    borderRadius: roundPx(28),
-    backgroundColor: '$primaryBlue',
-    alignItems: 'center',
-    justifyContent: 'center',
-    shadowColor: '$shadowColor',
-    shadowOffset: { width: 0, height: 4 },
-    shadowOpacity: 0.3,
-    shadowRadius: 6,
-    elevation: 6,
-  },
 });
