+++ src/components/bottomTabBar/view/bottomTabBarView.tsx
@@ -4,29 +4,24 @@ import { View, TouchableOpacity } from 'react-native';
 // Constants
 import { useDispatch } from 'react-redux';
 import { BottomTabBarProps } from '@react-navigation/bottom-tabs';
-import { useIntl } from 'react-intl';
+
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import { SheetManager } from 'react-native-actions-sheet';
+
 import ROUTES from '../../../constants/routeNames';
 
 // Styles
 import styles from './bottomTabBarStyles';
 import Icon, { IconContainer } from '../../icon';
 import { updateActiveBottomTab } from '../../../redux/actions/uiAction';
-import { useAppSelector } from '../../../hooks';
-import showLoginAlert from '../../../utils/showLoginAlert';
-import { SheetNames } from '../../../navigation/sheets';
 
 const BottomTabBarView = ({
   state: { routes, index },
   navigation,
   descriptors,
 }: BottomTabBarProps) => {
-  const intl = useIntl();
   const dispatch = useDispatch();
   const insets = useSafeAreaInsets();
-  const isLoggedIn = useAppSelector((state) => state.application.isLoggedIn);
-  const activeRoute = routes[index];
+
 
   useEffect(() => {
     dispatch(updateActiveBottomTab(routes[index].name));
@@ -80,33 +75,11 @@ const BottomTabBarView = ({
   });
 
   const _bottomPadding = insets.bottom || 16;
-  const _fabOffset = _bottomPadding + 70;
-
-  const _onCreatePress = () => {
-    if (!isLoggedIn) {
-      showLoginAlert({ intl });
-      return;
-    }
-
-    if (activeRoute.name === ROUTES.TABBAR.WAVES) {
-      SheetManager.show(SheetNames.QUICK_POST, {
-        payload: { mode: 'wave' },
-      });
-      return;
-    }
-
-    navigation.navigate(ROUTES.SCREENS.EDITOR, { key: 'editor_post' });
-  };
 
   return (
     <View style={[styles.wrapper, { paddingBottom: _bottomPadding }]}>
       {_tabButtons}
 
-      <View pointerEvents="box-none" style={[styles.fabWrapper, { bottom: _fabOffset }]}>
-        <TouchableOpacity style={styles.fabButton} onPress={_onCreatePress}>
-          <Icon iconType="MaterialCommunityIcons" name="pencil" color="#fff" size={24} />
-        </TouchableOpacity>
-      </View>
     </View>
   );
 };
