+++ src/components/buttons/index.ts
@@ -2,5 +2,6 @@ import CircularButton from './views/circularButtonView';
 import IconButton from './views/iconButtomView';
 import TextButton from './views/textButtonView';
 import SquareButton from './views/squareButton';
+import FabButton from './views/fabButton';
 
-export { CircularButton, IconButton, TextButton, SquareButton };
+export { CircularButton, IconButton, TextButton, SquareButton, FabButton };
