+++ src/screens/feed/screen/feedScreen.tsx
@@ -1,10 +1,13 @@
 import React, { Fragment, useMemo, useState } from 'react';
 import { View } from 'react-native';
 import get from 'lodash/get';
+import { useNavigation } from '@react-navigation/native';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import { useIntl } from 'react-intl';
 
 // Components
 import { gestureHandlerRootHOC } from 'react-native-gesture-handler';
-import { Header, TabbedPosts } from '../../../components';
+import { Header, TabbedPosts, FabButton } from '../../../components';
 
 // Styles
 import styles from './feedStyles';
@@ -18,8 +21,13 @@ import {
 import { useAppSelector } from '../../../hooks';
 import CommentsTabContent from '../../../components/profile/children/commentsTabContent';
 import { TabItem } from '../../../components/tabbedPosts/types/tabbedPosts.types';
+import ROUTES from '../../../constants/routeNames';
+import showLoginAlert from '../../../utils/showLoginAlert';
 
 const FeedScreen = () => {
+  const intl = useIntl();
+  const navigation = useNavigation();
+  const insets = useSafeAreaInsets();
   const isLoggedIn = useAppSelector((state) => state.application.isLoggedIn);
   const currentAccount = useAppSelector((state) => state.account.currentAccount);
 
@@ -44,10 +52,10 @@ const FeedScreen = () => {
     () =>
       feedFilters.map(
         (key: string) =>
-          ({
-            filterKey: key,
-            label: FEED_SCREEN_FILTER_MAP[key],
-          } as TabItem),
+        ({
+          filterKey: key,
+          label: FEED_SCREEN_FILTER_MAP[key],
+        } as TabItem),
       ),
     [feedFilters],
   );
@@ -72,6 +80,17 @@ const FeedScreen = () => {
     [feedFilters, currentAccount],
   );
 
+
+
+  const _onCreatePress = () => {
+    if (!isLoggedIn) {
+      showLoginAlert({ intl });
+      return;
+    }
+
+    navigation.navigate(ROUTES.SCREENS.EDITOR, { key: 'editor_post' });
+  };
+
   return (
     <Fragment>
       <Header showQR={true} showBoost={true} />
@@ -88,6 +107,7 @@ const FeedScreen = () => {
           />
         )}
       </View>
+      <FabButton bottomOffset={16} onPress={_onCreatePress} />
     </Fragment>
   );
 };
